refactored REST macro instances and companions out of RawRest
