+++ commons-core/src/main/scala/com/avsystem/commons/rest/RawRest.scala
@@ -103,50 +103,4 @@ object RawRest extends RawRpcCompanion[RawRest] {
       handleRequest(RestRequest(methodMeta.method, newHeaders, body))
     }
   }
-
-  trait ClientInstances[Real, I] extends RpcMacroInstances[Real] {
-    type Implicits = I
-    def metadata(implicits: I): RestMetadata[Real]
-    def asReal(implicits: I): AsRealRpc[Real]
-  }
-  trait ServerInstances[Real, I] extends RpcMacroInstances[Real] {
-    type Implicits = I
-    def metadata(implicits: I): RestMetadata[Real]
-    def asRaw(implicits: I): AsRawRpc[Real]
-  }
-  trait FullInstances[Real, I] extends RpcMacroInstances[Real] {
-    type Implicits = I
-    def metadata(implicits: I): RestMetadata[Real]
-    def asRawReal(implicits: I): AsRawRealRpc[Real]
-  }
-
-  // I have no idea why I can't just create one macro in `RpcMacroInstances` companion to rule them all
-  implicit def clientInstances[Real, I]: ClientInstances[Real, I] =
-  macro macros.rpc.RpcMacros.macroInstances[ClientInstances[Real, I], Real]
-  implicit def serverInstances[Real, I]: ServerInstances[Real, I] =
-  macro macros.rpc.RpcMacros.macroInstances[ServerInstances[Real, I], Real]
-  implicit def clientServerInstances[Real, I]: FullInstances[Real, I] =
-  macro macros.rpc.RpcMacros.macroInstances[FullInstances[Real, I], Real]
-
-  /** @see [[FullApiCompanion]]*/
-  abstract class ClientApiCompanion[Real, I](implicit instances: ClientInstances[Real, I]) { this: I =>
-    implicit final lazy val restMetadata: RestMetadata[Real] = instances.metadata(this)
-    implicit final lazy val restAsReal: AsRealRpc[Real] = instances.asReal(this)
-  }
-  /** @see [[FullApiCompanion]]*/
-  abstract class ServerApiCompanion[Real, I](implicit instances: ServerInstances[Real, I]) { this: I =>
-    implicit final lazy val restMetadata: RestMetadata[Real] = instances.metadata(this)
-    implicit final lazy val restAsRaw: AsRawRpc[Real] = instances.asRaw(this)
-  }
-  /**
-    * Base class for REST trait companions. Reduces boilerplate needed in order to define appropriate instances
-    * of `AsRawReal` and `RestMetadata` for given trait. The `I` type parameter lets you inject additional implicits
-    * into macro materialization of these instances, e.g. [[DefaultRestImplicits]].
-    * Usually, for even less boilerplate, this base class is extended by yet another abstract class which fixes
-    * the `I` type, e.g. [[RestApiCompanion]].
-    */
-  abstract class FullApiCompanion[Real, I](implicit instances: FullInstances[Real, I]) { this: I =>
-    implicit final lazy val restMetadata: RestMetadata[Real] = instances.metadata(this)
-    implicit final lazy val restAsRawReal: AsRawRealRpc[Real] = instances.asRawReal(this)
-  }
 }
